requirements.txt
langchain-core==1.0.4
langchain-openai==1.0.2
qdrant-client==1.7.0
huggingface-hub==1.1.2
deep-translator==1.11.4
langdetect==1.0.9
python-dotenv==1.0.0
